rename nats-messaging-client to just messaging
