+++ modules/node/package.json
@@ -20,7 +20,7 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand"
   },
   "dependencies": {
-    "@connext/nats-messaging-client": "0.0.5",
+    "@connext/messaging": "0.0.5",
     "@connext/types": "0.0.2",
     "@counterfactual/contracts": "0.1.4",
     "@counterfactual/firebase-client": "0.0.3",
