+++ ops/deploy-messaging.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -e
 
-package="`cat modules/nats-messaging-client/package.json | grep '"name":' | awk -F '"' '{print $4}'`"
+package="`cat modules/messaging/package.json | grep '"name":' | awk -F '"' '{print $4}'`"
 
 ########################################
 ## Run some sanity checks to make sure we're really ready to deploy
@@ -40,7 +40,7 @@ fi
 echo "Let's go"
 
 # edit this package's package.json to set new version number
-cd modules/nats-messaging-client
+cd modules/messaging
 mv package.json .package.json
 cat .package.json \
   | sed 's/"version": ".*"/"version": "'$target_version'"/' > package.json
@@ -60,8 +60,7 @@ done
 
 cd ../..
 git add .
-git add modules/nats-messaging-client modules/node
 git commit --allow-empty -m "Publish package: $package@$target_version"
-git tag nats-client-$target_version
+git tag messaging-$target_version
 git push origin HEAD --no-verify
-git push origin nats-client-$target_version --no-verify
+git push origin messaging-$target_version --no-verify
